Three different projects with the implementation of the following challenges.
test-palindrome
Write an efficient algorithm to check if a string is a palindrome. A string is a palindrome if the string matches the reverse of string. 

test-k-complementary
Write an efficient algorithm to find K-complementary pairs in a given array of integers.
Given Array A, pair (i, j) is K- complementary if K = A[i] + A[j];

test-term-frequency
Tf/idf (term frequency / inverse document frequency) is an statistic that reflects the importance of a term T in a document D (or the relevance of a document for a searched term) relative to a document set S.